pyproject.toml
[tool.poetry]
name = "scalarizing"
version = "0.1.0"
description = "Scalarizing"
authors = ["Bogdan Gulowaty <b.gulowaty@gmail.com>"]

packages = [
    { include = "scalarizing" },
]

[tool.pytest.ini_options]
python_files = "*.py"
norecursedirs=["libs"]

[tool.poetry.dependencies]
python = "^3.8"
python-box = {extras = ["all"], version = "^5.4.1"}
x-to-rules = {path = "libs/rules", develop = true}
pandas = "1.4.0"
sklearn = "^0.0"
pymoo = "^0.6.0"
imblearn = "^0.0"
toolz = "^0.11.2"
loky = "^3.1.0"
mlxtend = "^0.21.0"
wrapt = "^1.14.1"
ray = {extras = ["default"], version = "^2.0.0"}
datasets = "^2.6.1"
DESlib = "^0.3.5"
perceptron = "^1.1.0"

[tool.poetry.group.dev.dependencies]
bokeh = "^2.4.3"
mlflow = "^1.23.1"
jupyterlab = "^3.2.8"
distributed = "^2022.10.0"
dask = {extras = ["distributed"], version = "^2022.6.1"}


[build-system]
requires = ["poetry-core>=1.1.0a6"]
